_archived/requirements.txt
bokeh==3.4.1
cairocffi==1.6.1
cffi==1.16.0
click==8.1.7
contextlib2==21.6.0
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
find-libpython==0.4.0
fonttools==4.51.0
future==1.0.0
importlib-resources==6.4.0
inspyred==1.0.2
jinja2==3.1.3
kiwisolver==1.4.5
lfpykit==0.5.1
lxml==5.2.1
markupsafe==2.1.5
matplotlib==3.8.4
matplotlib-scalebar==0.8.1
meautility==1.5.1
netpyne==1.0.6
neuron==8.2.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
schema==0.7.5
scipy==1.13.0
six==1.16.0
tinycss2==1.2.1
tornado==6.4
tzdata==2024.1
webencodings==0.5.1
xyzservices==2024.4.0
zipp==3.18.1
